drop dnnlib utils
